@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0",
+    "@angular/animations": "^14.2.1",
     "@angular/cdk": "^13.1.3",
-    "@angular/common": "~13.0.0",
-    "@angular/compiler": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/forms": "~13.0.0",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/common": "^14.2.1",
+    "@angular/compiler": "^14.2.1",
+    "@angular/core": "^14.2.1",
+    "@angular/forms": "^14.2.1",
+    "@angular/platform-browser": "^14.2.1",
+    "@angular/platform-browser-dynamic": "^14.2.1",
+    "@angular/router": "^14.2.1",
     "@fortawesome/fontawesome-free": "^5.15.4",
     "@types/jquery": "^3.5.14",
     "@types/lodash": "^4.14.182",
@@ -39,8 +39,8 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.2.2",
-    "@angular/cli": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.0",
+    "@angular/cli": "^14.2.2",
+    "@angular/compiler-cli": "^14.2.1",
     "@types/aos": "^3.0.4",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
@@ -54,6 +54,6 @@
     "node-sass": "^7.0.3",
     "postcss": "^8.4.14",
     "sass-loader": "^13.0.2",
-    "typescript": "~4.4.3"
+    "typescript": "~4.8.3"
   }
 }
